infra/mcp-runner/requirements.txt
elasticsearch>=8.11.0
python-dotenv>=1.0.0
requests>=2.31.0
docker>=7.0.0
